+++ core/boot.js
@@ -90,9 +90,9 @@ async function initCore({ghostServer, config}) {
     await settings.syncEmailSettings(config.get('hostSettings:emailVerification:verified'));
     debug('End: settings');
 
-    // The URLService is a core part of Ghost, which depends on models. It needs moving from the frontend to make this clear.
+    // The URLService is a core part of Ghost, which depends on models.
     debug('Begin: Url Service');
-    const urlService = require('./frontend/services/url');
+    const urlService = require('./server/services/url');
     // Note: there is no await here, we do not wait for the url service to finish
     // We can return, but the site will remain in (the shared, not global) maintenance mode until this finishes
     // This is managed on request: https://github.com/TryGhost/Ghost/blob/main/core/server/web/shared/middlewares/maintenance.js#L13
