+++ core/server/services/route-settings/route-settings.js
@@ -3,7 +3,7 @@ const moment = require('moment-timezone');
 const fs = require('fs-extra');
 const path = require('path');
 const crypto = require('crypto');
-const urlService = require('../../../frontend/services/url');
+const urlService = require('../url');
 
 const debug = require('@tryghost/debug')('services:route-settings');
 const errors = require('@tryghost/errors');
