+++ core/server/services/slack.js
@@ -4,7 +4,7 @@ const logging = require('@tryghost/logging');
 const request = require('@tryghost/request');
 const {blogIcon} = require('../lib/image');
 const urlUtils = require('../../shared/url-utils');
-const urlService = require('../../frontend/services/url');
+const urlService = require('./url');
 const settingsCache = require('../../shared/settings-cache');
 const schema = require('../data/schema').checks;
 const moment = require('moment');
