+++ core/server/services/url/Resource.js
