+++ core/server/services/url/Urls.js
@@ -5,7 +5,7 @@ const logging = require('@tryghost/logging');
 const errors = require('@tryghost/errors');
 
 // This emits its own url added/removed events
-const events = require('../../../server/lib/common/events');
+const events = require('../../lib/common/events');
 
 /**
  * This class keeps track of all urls in the system.
