+++ core/server/services/xmlrpc.js
@@ -1,7 +1,7 @@
 const _ = require('lodash');
 const xml = require('xml');
 const config = require('../../shared/config');
-const urlService = require('../../frontend/services/url');
+const urlService = require('./url');
 const errors = require('@tryghost/errors');
 const tpl = require('@tryghost/tpl');
 const logging = require('@tryghost/logging');
