+++ core/server/web/site/app.js
@@ -9,7 +9,7 @@ const errors = require('@tryghost/errors');
 const config = require('../../../shared/config');
 const constants = require('@tryghost/constants');
 const storage = require('../../adapters/storage');
-const urlService = require('../../../frontend/services/url');
+const urlService = require('../../services/url');
 const urlUtils = require('../../../shared/url-utils');
 const sitemapHandler = require('../../../frontend/services/sitemap/handler');
 const appService = require('../../../frontend/services/apps');
