+++ test/regression/models/model_posts.test.js
@@ -6,7 +6,7 @@ const moment = require('moment');
 const _ = require('lodash');
 const Promise = require('bluebird');
 const {sequence} = require('@tryghost/promise');
-const urlService = require('../../../core/frontend/services/url');
+const urlService = require('../../../core/server/services/url');
 const ghostBookshelf = require('../../../core/server/models/base');
 const models = require('../../../core/server/models');
 const db = require('../../../core/server/data/db');
