+++ test/unit/frontend/apps/amp/router.test.js
@@ -3,7 +3,7 @@ const should = require('should');
 const sinon = require('sinon');
 const path = require('path');
 const ampController = require('../../../../../core/frontend/apps/amp/lib/router');
-const urlService = require('../../../../../core/frontend/services/url');
+const urlService = require('../../../../../core/server/services/url');
 const helpers = require('../../../../../core/frontend/services/routing/helpers');
 const testUtils = require('../../../../utils');
 const configUtils = require('../../../../utils/configUtils');
