+++ test/unit/frontend/helpers/author.test.js
@@ -1,7 +1,7 @@
 const should = require('should');
 const sinon = require('sinon');
 const testUtils = require('../../../utils');
-const urlService = require('../../../../core/frontend/services/url');
+const urlService = require('../../../../core/server/services/url');
 const authorHelper = require('../../../../core/frontend/helpers/author');
 
 describe('{{author}} helper', function () {
