+++ test/unit/frontend/helpers/url.test.js
@@ -6,7 +6,7 @@ const testUtils = require('../../../utils');
 const configUtils = require('../../../utils/configUtils');
 const markdownToMobiledoc = require('../../../utils/fixtures/data-generator').markdownToMobiledoc;
 const url = require('../../../../core/frontend/helpers/url');
-const urlService = require('../../../../core/frontend/services/url');
+const urlService = require('../../../../core/server/services/url');
 const api = require('../../../../core/server/api');
 
 describe('{{url}} helper', function () {
