+++ test/unit/frontend/meta/url.test.js
@@ -1,7 +1,7 @@
 const should = require('should');
 const sinon = require('sinon');
 const urlUtils = require('../../../../core/shared/url-utils');
-const urlService = require('../../../../core/frontend/services/url');
+const urlService = require('../../../../core/server/services/url');
 const getUrl = require('../../../../core/frontend/meta/url');
 const testUtils = require('../../../utils');
