+++ test/unit/frontend/services/url/Queue.test.js
@@ -2,7 +2,7 @@ const _ = require('lodash');
 const should = require('should');
 const sinon = require('sinon');
 const logging = require('@tryghost/logging');
-const Queue = require('../../../../../core/frontend/services/url/Queue');
+const Queue = require('../../../../../core/server/services/url/Queue');
 
 describe('Unit: services/url/Queue', function () {
     let queue;
