+++ test/unit/frontend/services/url/Urls.test.js
@@ -1,7 +1,7 @@
 const should = require('should');
 const sinon = require('sinon');
 const events = require('../../../../../core/server/lib/common/events');
-const Urls = require('../../../../../core/frontend/services/url/Urls');
+const Urls = require('../../../../../core/server/services/url/Urls');
 
 describe('Unit: services/url/Urls', function () {
     let urls;
