+++ test/utils/e2e-utils.js
@@ -14,7 +14,7 @@ const config = require('../../core/shared/config');
 const boot = require('../../core/boot');
 const db = require('../../core/server/data/db');
 const models = require('../../core/server/models');
-const urlService = require('../../core/frontend/services/url');
+const urlService = require('../../core/server/services/url');
 const settingsService = require('../../core/server/services/settings');
 const routeSettingsService = require('../../core/server/services/route-settings');
 const web = require('../../core/server/web');
